Reorganizing Extractors
ECM-based extractors are now reorganized, so that it is easier to make
modifications and improvements to individual components without having to deal
with a large file. All tests are passing as well.
